sprint1/step1 - renaming constructor to ingredients
