+++ src/components/burger-constructor/components/basket-order/burger-constructor-order.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './basket-order.module.css';
+import styles from './burger-constructor-order.module.css';
 import {Button} from '@ya.praktikum/react-developer-burger-ui-components';
 import {MoneyCounter} from '../../../common/money-counter/money-counter';
 
@@ -7,7 +7,7 @@ interface IBasketOrderProps {
     sum: number;
 }
 
-export function BasketOrder({sum}: IBasketOrderProps) {
+export function BurgerConstructorOrder({sum}: IBasketOrderProps) {
     return (
         <div className={`mr-4 ${styles.main}`}>
             <div className='mr-10'><MoneyCounter sum={sum} big/></div>
