+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -10,7 +10,3 @@
     overflow-y: scroll;
     flex-direction: column;
 }
-
-.sum {
-    align-self: flex-end;
-}
